Fixes to OEM driver

+++ src/EC_OEM.cpp
@@ -82,7 +82,7 @@ void EC_OEM::initializeInstanceVariables(uint8_t pin, uint8_t addr_)
 void EC_OEM::init(bool led_, bool hibernate_, uint8_t int_CTRL)
 {
   this->wire->beginTransmission(device_addr);
-  this->wire->write(device_type);
+  this->wire->write((uint8_t) device_type);
   if( this->wire->endTransmission() == 0 )
   {
     if(getDeviceType() == EC_OEM_DEVICE )
